finishes sdm tests
