+++ menpo/fitmultilevel/sdm/test/sdm_test.py
@@ -1,13 +1,12 @@
 from mock import patch
-import numpy as np
-from numpy.testing import assert_allclose
 from nose.tools import raises
 from StringIO import StringIO
+from nose.plugins.attrib import attr
 
+import numpy as np
 import menpo.io as mio
 from menpo.landmark import labeller, ibug_68_trimesh
 from menpo.fitmultilevel.sdm import SDMTrainer, SDAAMTrainer, SDCLMTrainer
-from menpo.shape import PointCloud
 from menpo.transform.modeldriven import OrthoMDTransform
 from menpo.transform.homogeneous import AlignmentSimilarity
 from menpo.fitmultilevel.featurefunctions import sparse_hog
@@ -18,7 +17,7 @@ from menpo.transform import PiecewiseAffine
 from menpo.fitmultilevel.aam import AAMBuilder
 from menpo.fitmultilevel.clm import CLMBuilder
 from menpo.model.modelinstance import OrthoPDM
-
+from menpo.shape import PointCloud
 
 initial_shape = []
 initial_shape.append(PointCloud(np.array([[150.9737801, 1.85331141],
@@ -307,3 +306,247 @@ for i in range(4):
     if im.n_channels == 3:
         im = im.as_greyscale(mode='luminosity')
     training_images.append(im)
+
+# Seed the random number generator
+np.random.seed(seed=1000)
+
+# build sdms
+sdm1 = SDMTrainer(regression_type=mlr_svd,
+                  regression_features=sparse_hog,
+                  patch_shape=(16, 16),
+                  feature_type=None,
+                  normalization_diagonal=150,
+                  n_levels=2,
+                  downscale=1.3,
+                  pyramid_on_features=True,
+                  noise_std=0.04,
+                  rotation=False,
+                  n_perturbations=2,
+                  interpolator='scipy').train(training_images, group='PTS')
+
+aam = AAMBuilder(feature_type=sparse_hog,
+                 transform=PiecewiseAffine,
+                 trilist=training_images[0].landmarks['ibug_68_trimesh'].
+                 lms.trilist,
+                 normalization_diagonal=150,
+                 n_levels=3,
+                 downscale=1.2,
+                 scaled_shape_models=False,
+                 pyramid_on_features=True,
+                 max_shape_components=None,
+                 max_appearance_components=3,
+                 boundary=3,
+                 interpolator='scipy').build(training_images, group='PTS')
+
+sdm2 = SDAAMTrainer(aam,
+                    regression_type=mlr,
+                    regression_features=weights,
+                    noise_std=0.04,
+                    rotation=False,
+                    n_perturbations=1,
+                    update='compositional',
+                    md_transform=OrthoMDTransform,
+                    global_transform=AlignmentSimilarity,
+                    n_shape=25,
+                    n_appearance=None).train(training_images, group='PTS')
+
+clm = CLMBuilder(classifier_type=linear_svm_lr,
+                 feature_type=[sparse_hog],
+                 normalization_diagonal=100,
+                 patch_shape=(5, 5),
+                 n_levels=1,
+                 downscale=1.1,
+                 scaled_shape_models=True,
+                 pyramid_on_features=True,
+                 max_shape_components=25,
+                 boundary=3,
+                 interpolator='scipy').build(training_images, group='PTS')
+
+sdm3 = SDCLMTrainer(clm,
+                    regression_type=mlr,
+                    noise_std=0.04,
+                    rotation=False,
+                    n_perturbations=1,
+                    pdm_transform=OrthoPDM,
+                    global_transform=AlignmentSimilarity,
+                    n_shape=None).train(training_images, group='PTS')
+
+
+@raises(ValueError)
+def test_feature_type_exception():
+    sdm = SDMTrainer(feature_type=['igo', sparse_hog],
+                     n_levels=3).train(training_images, group='PTS')
+
+
+@raises(ValueError)
+def test_feature_type_with_pyramid_on_features_exception():
+    sdm = SDMTrainer(feature_type=['igo', sparse_hog, 'hog'],
+                     n_levels=3,
+                     pyramid_on_features=True).train(training_images,
+                                                     group='PTS')
+
+
+@raises(ValueError)
+def test_regression_features_sdmtrainer_exception_1():
+    sdm = SDMTrainer(n_levels=2, regression_features=[None, None, None]).\
+        train(training_images, group='PTS')
+
+
+@raises(ValueError)
+def test_regression_features_sdmtrainer_exception_2():
+    sdm = SDMTrainer(n_levels=3, regression_features=[None, sparse_hog, 1]).\
+        train(training_images, group='PTS')
+
+
+@raises(ValueError)
+def test_regression_features_sdaamtrainer_exception_1():
+    sdm = SDAAMTrainer(aam, regression_features=[None, sparse_hog]).\
+        train(training_images, group='PTS')
+
+
+@raises(ValueError)
+def test_regression_features_sdaamtrainer_exception_2():
+    sdm = SDAAMTrainer(aam, regression_features=7).\
+        train(training_images, group='PTS')
+
+
+@raises(ValueError)
+def test_n_levels_exception():
+    sdm = SDMTrainer(n_levels=0).train(training_images, group='PTS')
+
+
+@raises(ValueError)
+def test_downscale_exception():
+    sdm = SDMTrainer(downscale=1).train(training_images,
+                                        group='PTS')
+    assert (aam.downscale == 1)
+    sdm = SDMTrainer(downscale=0).train(training_images,
+                                        group='PTS')
+
+
+@raises(ValueError)
+def test_n_perturbations_exception():
+    sdm = SDAAMTrainer(aam, n_perturbations=-10).train(training_images,
+                                                       group='PTS')
+
+
+@patch('sys.stdout', new_callable=StringIO)
+def test_verbose_mock(mock_stdout):
+    sdm = SDMTrainer(regression_type=mlr_svd,
+                     regression_features=sparse_hog,
+                     patch_shape=(16, 16),
+                     feature_type=None,
+                     normalization_diagonal=150,
+                     n_levels=1,
+                     downscale=1.3,
+                     pyramid_on_features=True,
+                     noise_std=0.04,
+                     rotation=False,
+                     n_perturbations=2,
+                     interpolator='scipy').train(training_images, group='PTS',
+                                                 verbose=True)
+
+
+@patch('sys.stdout', new_callable=StringIO)
+def test_str_mock(mock_stdout):
+    print sdm1
+    print sdm2
+    print sdm3
+    
+    
+def test_sdm_1():
+    assert (sdm1._n_training_images == 4)
+    assert (sdm1.n_levels == 2)
+    assert (sdm1.downscale == 1.3)
+    assert (sdm1.feature_type[0] is None)
+    assert (sdm1.interpolator == 'scipy')
+    assert (sdm1.algorithm == 'SDM-Non-Parametric')
+    assert sdm1.pyramid_on_features
+    assert (sdm1._fitters[0].algorithm == sdm1._fitters[1].algorithm ==
+            'Non-Parametric')
+    assert (sdm1._fitters[0].regressor.__name__ ==
+            sdm1._fitters[1].regressor.__name__ == 'mlr_svd_fitting')
+
+
+def test_sdm_2():
+    assert (sdm2._n_training_images == 4)
+    assert (sdm2.n_levels == 3)
+    assert (sdm2.downscale == 1.2)
+    assert (sdm2.interpolator == 'scipy')
+    assert (sdm2.algorithm == 'SD-AAM-Parametric')
+    assert sdm2.pyramid_on_features
+    assert (sdm2._fitters[0].algorithm == sdm2._fitters[1].algorithm ==
+            sdm2._fitters[2].algorithm == 'Parametric')
+    assert (sdm2._fitters[0].regressor.__name__ ==
+            sdm2._fitters[1].regressor.__name__ ==
+            sdm2._fitters[2].regressor.__name__ == 'mlr_fitting')
+
+
+def test_sdm_3():
+    assert (sdm3._n_training_images == 4)
+    assert (sdm3.n_levels == 1)
+    assert (sdm3.downscale == 1.1)
+    assert (sdm3.interpolator == 'scipy')
+    assert (sdm3.algorithm == 'SD-CLM-Semi-Parametric')
+    assert sdm3.pyramid_on_features
+    assert (sdm3._fitters[0].algorithm == 'Semi-Parametric')
+    assert (sdm3._fitters[0].regressor.__name__ == 'mlr_fitting')
+
+
+def test_pertrurb_shape():
+    s = sdm1.perturb_shape(training_images[0].landmarks['PTS'].lms,
+                           noise_std=0.08, rotation=False)
+    assert (s.n_dims == 2)
+    assert (s.n_landmark_groups == 0)
+    assert (s.n_points == 68)
+
+
+def test_obtain_shape_from_bb():
+    s = sdm2.obtain_shape_from_bb(np.array([[26, 49], [350, 400]]))
+    assert ((np.around(s.points) == np.around(initial_shape[3].points)).
+            all())
+    assert (s.n_dims == 2)
+    assert (s.n_landmark_groups == 0)
+    assert (s.n_points == 68)
+
+
+@raises(ValueError)
+def test_max_iters_exception():
+    sdm3.fit(training_images[0], initial_shape[0], max_iters=[10, 20, 30, 40])
+
+
+def sdm_helper(sdm, im_number, max_iters, initial_error, final_error,
+               error_type, str_flag):
+    fitting_result = sdm.fit(
+        training_images[im_number], initial_shape[im_number],
+        gt_shape=training_images[im_number].landmarks['PTS'].lms,
+        max_iters=max_iters, error_type=error_type)
+    if str_flag is True:
+        ie = str(fitting_result.initial_error)
+        fe = str(fitting_result.final_error)
+        assert (ie[0:5] == initial_error)
+        assert (fe[0:5] == final_error)
+    else:
+        assert (np.around(fitting_result.initial_error, 5) == initial_error)
+        assert (np.around(fitting_result.final_error, 5) == final_error)
+
+
+@attr('fuzzy')
+def test_sdm1_fit():
+    sdm_helper(sdm1, 0, None, 0.084, 0.0352, 'me_norm', False)
+    sdm_helper(sdm1, 1, 10, 8.66397, 2.13049, 'me', False)
+    sdm_helper(sdm1, 2, [5, 10], 13.94481, 9.04642, 'rmse', False)
+
+
+@attr('fuzzy')
+def test_sdm2_fit():
+    sdm_helper(sdm2, 0, None, 0.08166, 5.17755, 'me_norm', False)
+    sdm_helper(sdm2, 1, 10, 7.8825, 140.20838, 'me', False)
+    sdm_helper(sdm2, 2, [5, 10, 15], 11.43388, 369.90081, 'rmse', False)
+
+
+@attr('fuzzy')
+def test_sdm3_fit():
+    sdm_helper(sdm3, 0, None, '0.079', '4.891', 'me_norm', True)
+    sdm_helper(sdm3, 1, None, '7.294', '1.492', 'me', True)
+    sdm_helper(sdm3, 2, None, '10.69', '1.382', 'rmse', True)
